Override DefaultDependencies on ipset service unit
The current ipset service unit sets `Before=network-pre.target` as it is
needed before the network to help ensure that the firewall is up and
running before anyone might actually try to connect. It has no setting
for `DefaultDependencies` which means it takes the defaults, one of
which is `After=sysinit.target`.

There are some cases where this can cause a dependency cycle with other
units that want to start early in the boot process between
`networking.service` and `sysinit.target` (for example, cloud-init).
This results in one of the offending units being removed and risks
startup continuing without ipsets being configured, potentially leaving
a firewall open and a host at risk.

The ipset service can safely be run before `sysinit.target` by setting
`DefaultDependencies=no` with a couple of additional dependencies to
handle some of the other implied settings that are removed as a result.

Fixes #63